Switch to on-demand loading and memory-mapped files where available.
GCW Zero: Report decompression progress and which files are being loaded/saved during startup and shutdown.

+++ source/memory.h
@@ -277,7 +277,7 @@ extern CPU_ALERT_TYPE dma_transfer(DMA_TRANSFER_TYPE *dma);
 extern u8 *memory_region(u32 address, u32 *memory_limit);
 extern s32 load_bios(char *name);
 extern ssize_t load_gamepak(char *file_path);
-extern u8 *load_gamepak_page(u32 physical_index);
+extern u8 *load_gamepak_page(u16 physical_index);
 extern u32 load_backup();
 extern void init_memory();
 extern void init_gamepak_buffer();
